refactor: phases 2.1 and 3 - externalize config and split tool handlers
Phase 2.1: Externalize Model Configuration
- Add config/models.yaml with all model definitions
- Add services/ai_config.go with thread-safe config loading
- Use ZETTELGARDEN_CONFIG_PATH env var for config path
- Replace hardcoded ValidChatModels with YAML-based config
- Add sync.Once for thread-safe lazy initialization
- Add warning log when model pricing not found

Phase 3: Split Tool Handlers by Domain (memory_tools proof of concept)
- Add services/featureflags/ package for feature flag support
- Create services/tools/memory/ domain package
- Split memory_tools with feature flag: MEMORY_TOOLS_V2
- Add comprehensive integration tests for both paths
- Fix type inconsistency (userID: uint → int)

Benefits:
- Add/update models via YAML without code changes
- Thread-safe config loading with proper error handling
- Feature flags enable gradual rollout of domain splits
- Pattern established for remaining 7 domains

+++ go-backend/handlers/memory.go
@@ -85,7 +85,7 @@ func GetUserMemory(db *sql.DB, userID int) (string, error) {
 }
 
 // UpdateUserMemory is a wrapper for services.UpdateUserMemory for backward compatibility
-func UpdateUserMemory(db *sql.DB, userID uint, memory string) error {
+func UpdateUserMemory(db *sql.DB, userID int, memory string) error {
 	return services.UpdateUserMemory(db, userID, memory)
 }
 
@@ -124,7 +124,7 @@ func (s *Handler) UpdateUserMemoryRoute(w http.ResponseWriter, r *http.Request) 
 	}
 
 	// Update memory in database
-	err = UpdateUserMemory(s.DB, uint(userID), requestData.Memory)
+	err = UpdateUserMemory(s.DB, userID, requestData.Memory)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -192,7 +192,7 @@ Your task is to produce a new, superior, and more compact version of the entire 
 		return "", fmt.Errorf("no response from AI")
 	}
 
-	err = UpdateUserMemory(db, uint(userID), response.Choices[0].Message.Content)
+	err = UpdateUserMemory(db, int(userID), response.Choices[0].Message.Content)
 	if err != nil {
 		return "", err
 	}
